+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * [HexaPDF::Layout::Frame::FitResult#draw] to allow drawing at an offset
 * [HexaPDF::Layout::Box#fit] to delegate the actual content fitting to the
   `#fit_content` method
+* [HexaPDF::Document::Layout#box] to allow using the block as drawing block for
+  the base box class
 
 
 ## 0.35.1 - 2024-01-11